Part #15 (Cart Index, Not Found, Remove Item)

+++ src/VirtualShop/VirtualShop.Web/Models/CartHeaderViewModel.cs
@@ -6,6 +6,6 @@ namespace VirtualShop.Web.Models
         public string UserId { get; set; } = "";
         public string CouponCode { get; set; } = "";
 
-        public double TotalAmount { get; set; } = 0.00d;
+        public decimal TotalAmount { get; set; } = 0.00M;
     }
 }
